mri2ct/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
monai[all]>=1.3.0
torchio>=0.19.0
SimpleITK>=2.3.0
antspyx>=0.4.0
hydra-core>=1.3.0
omegaconf>=2.3.0
wandb>=0.15.0
matplotlib>=3.7.0
panel>=1.3.0
gradio>=4.0.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
scipy>=1.11.0
nibabel>=5.1.0
pandas>=2.0.0
tqdm>=4.65.0
einops>=0.7.0
timm>=0.9.0
lpips>=0.1.4
pillow>=10.0.0
pydicom>=2.4.0
